Revert "[CORL-382] Allow moderators to ban a commenter from the comment stream (#2378)"
This reverts commit e77103d87243d4c8934e639f7102e0680ea442f7.
