+++ src/core/client/stream/tabs/Comments/Comment/ModerationDropdown/ModerationDropdownContainer.css
@@ -10,13 +10,3 @@
   color: var(--palette-primary-dark);
   font-weight: var(--font-weight-bold);
 }
-.banned {
-  color: var(--palette-error-dark);
-  font-weight: var(--font-weight-bold);
-}
-.banIcon {
-  width: 18px;
-  height: 14px;
-  text-align: center;
-  line-height: 12px;
-}
