+++ src/core/client/stream/tabs/Comments/Comment/ModerationDropdown/ModerationDropdownContainer.tsx
@@ -1,54 +1,153 @@
-import React, { FunctionComponent, useCallback, useState } from "react";
+import { Localized } from "fluent-react/compat";
+import React, { FunctionComponent, useCallback } from "react";
 import { graphql } from "react-relay";
 
-import { withFragmentContainer } from "coral-framework/lib/relay";
+import { useMutation, withFragmentContainer } from "coral-framework/lib/relay";
 import { ModerationDropdownContainer_comment } from "coral-stream/__generated__/ModerationDropdownContainer_comment.graphql";
 import { ModerationDropdownContainer_story } from "coral-stream/__generated__/ModerationDropdownContainer_story.graphql";
-import { ModerationDropdownContainer_viewer } from "coral-stream/__generated__/ModerationDropdownContainer_viewer.graphql";
-import { Dropdown } from "coral-ui/components";
+import {
+  Dropdown,
+  DropdownButton,
+  DropdownDivider,
+  Icon,
+} from "coral-ui/components";
 
-import UserBanPopoverContainer from "../UserBanPopover/UserBanPopoverContainer";
-import ModerationActionsContainer from "./ModerationActionsContainer";
+import ApproveCommentMutation from "./ApproveCommentMutation";
+import FeatureCommentMutation from "./FeatureCommentMutation";
+import RejectCommentMutation from "./RejectCommentMutation";
+import UnfeatureCommentMutation from "./UnfeatureCommentMutation";
 
-type View = "MODERATE" | "BAN";
+import styles from "./ModerationDropdownContainer.css";
 
 interface Props {
   comment: ModerationDropdownContainer_comment;
   story: ModerationDropdownContainer_story;
-  viewer: ModerationDropdownContainer_viewer | null;
   onDismiss: () => void;
-  scheduleUpdate: () => void;
 }
 
 const ModerationDropdownContainer: FunctionComponent<Props> = ({
   comment,
   story,
-  viewer,
   onDismiss,
-  scheduleUpdate,
 }) => {
-  const [view, setView] = useState<View>("MODERATE");
-  const onBan = useCallback(() => {
-    setView("BAN");
-    scheduleUpdate();
-  }, [setView, scheduleUpdate]);
+  const approve = useMutation(ApproveCommentMutation);
+  const feature = useMutation(FeatureCommentMutation);
+  const unfeature = useMutation(UnfeatureCommentMutation);
+  const reject = useMutation(RejectCommentMutation);
+
+  const onApprove = useCallback(() => {
+    approve({ commentID: comment.id, commentRevisionID: comment.revision.id });
+  }, [approve, comment]);
+  const onReject = useCallback(
+    () =>
+      reject({ commentID: comment.id, commentRevisionID: comment.revision.id }),
+    [approve, comment]
+  );
+  const onFeature = useCallback(() => {
+    feature({
+      storyID: story.id,
+      commentID: comment.id,
+      commentRevisionID: comment.revision.id,
+    });
+    onDismiss();
+  }, [feature, comment]);
+  const onUnfeature = useCallback(() => {
+    unfeature({
+      commentID: comment.id,
+      storyID: story.id,
+    });
+    onDismiss();
+  }, [unfeature, comment]);
+
+  const approved = comment.status === "APPROVED";
+  const rejected = comment.status === "REJECTED";
+  const featured = comment.tags.some(t => t.code === "FEATURED");
 
   return (
-    <div>
-      {view === "MODERATE" ? (
-        <Dropdown>
-          <ModerationActionsContainer
-            comment={comment}
-            story={story}
-            viewer={viewer}
-            onDismiss={onDismiss}
-            onBan={onBan}
-          />
-        </Dropdown>
+    <Dropdown>
+      {featured ? (
+        <Localized id="comments-moderationDropdown-unfeature">
+          <DropdownButton
+            icon={
+              <Icon className={styles.featured} size="md">
+                star
+              </Icon>
+            }
+            className={styles.featured}
+            onClick={onUnfeature}
+          >
+            Un-Feature
+          </DropdownButton>
+        </Localized>
       ) : (
-        <UserBanPopoverContainer user={comment.author!} onDismiss={onDismiss} />
+        <Localized id="comments-moderationDropdown-feature">
+          <DropdownButton
+            icon={<Icon size="md">star_border</Icon>}
+            onClick={onFeature}
+          >
+            Feature
+          </DropdownButton>
+        </Localized>
       )}
-    </div>
+      {approved ? (
+        <Localized id="comments-moderationDropdown-approved">
+          <DropdownButton
+            icon={
+              <Icon className={styles.approved} size="md">
+                check
+              </Icon>
+            }
+            className={styles.approved}
+            disabled
+          >
+            Approved
+          </DropdownButton>
+        </Localized>
+      ) : (
+        <Localized id="comments-moderationDropdown-approve">
+          <DropdownButton
+            icon={<Icon size="md">check</Icon>}
+            onClick={onApprove}
+          >
+            Approve
+          </DropdownButton>
+        </Localized>
+      )}
+      {rejected ? (
+        <Localized id="comments-moderationDropdown-rejected">
+          <DropdownButton
+            icon={
+              <Icon className={styles.rejected} size="md">
+                close
+              </Icon>
+            }
+            className={styles.rejected}
+            disabled
+          >
+            Rejected
+          </DropdownButton>
+        </Localized>
+      ) : (
+        <Localized id="comments-moderationDropdown-reject">
+          <DropdownButton
+            icon={<Icon size="md">close</Icon>}
+            onClick={onReject}
+          >
+            Reject
+          </DropdownButton>
+        </Localized>
+      )}
+      <DropdownDivider />
+      <Localized id="comments-moderationDropdown-goToModerate">
+        <DropdownButton
+          href={`/admin/moderate/comment/${comment.id}`}
+          target="_blank"
+          anchor
+        >
+          Go to Moderate
+        </DropdownButton>
+      </Localized>
+    </Dropdown>
   );
 };
 
@@ -56,11 +155,6 @@ const enhanced = withFragmentContainer<Props>({
   comment: graphql`
     fragment ModerationDropdownContainer_comment on Comment {
       id
-      author {
-        id
-        username
-        ...UserBanPopoverContainer_user
-      }
       revision {
         id
       }
@@ -68,18 +162,11 @@ const enhanced = withFragmentContainer<Props>({
       tags {
         code
       }
-      ...ModerationActionsContainer_comment
     }
   `,
   story: graphql`
     fragment ModerationDropdownContainer_story on Story {
       id
-      ...ModerationActionsContainer_story
-    }
-  `,
-  viewer: graphql`
-    fragment ModerationDropdownContainer_viewer on User {
-      ...ModerationActionsContainer_viewer
     }
   `,
 })(ModerationDropdownContainer);
