--- src/core/client/stream/tabs/Comments/Comment/UserBanPopover/UserBanPopoverContainer.css
@@ -1,4 +1,0 @@
-.root {
-  width: 280px;
-  max-width: 80vw;
-}
