+++ src/core/client/stream/test/fixtures.ts
@@ -90,15 +90,6 @@ export const baseUser = createFixture<GQLUser>({
   createdAt: "2018-02-06T18:24:00.000Z",
   status: {
     current: [GQLUSER_STATUS.ACTIVE],
-    ban: {
-      active: false,
-      history: [],
-    },
-    suspension: {
-      active: false,
-      until: null,
-      history: [],
-    },
   },
   ignoredUsers: [],
   comments: {
